Conflicts:
	src/HappyEngine/code/HappyPCH.h
	src/HappyEngine/code/HappyTypes.h
	src/HappyEngine/code/OpenGL.h
	src/HappyEngine/code/containers/ObjectFactory.h
	src/HappyEngine/code/content/BinObjLoader.cpp
	src/HappyEngine/code/content/BinObjLoader.h
	src/HappyEngine/code/content/ContentManager.cpp
	src/HappyEngine/code/content/ContentManager.h
	src/HappyEngine/code/content/FontLoader.cpp
	src/HappyEngine/code/content/FontLoader.h
	src/HappyEngine/code/content/IModelLoader.h
	src/HappyEngine/code/content/MaterialLoader.cpp
	src/HappyEngine/code/content/MaterialLoader.h
	src/HappyEngine/code/content/ModelLoader.cpp
	src/HappyEngine/code/content/ModelLoader.h
	src/HappyEngine/code/content/ObjLoader.cpp
	src/HappyEngine/code/content/ObjLoader.h
	src/HappyEngine/code/content/PhysicsShapeLoader.cpp
	src/HappyEngine/code/content/ShaderLoader.cpp
	src/HappyEngine/code/content/TextureLoader.h
	src/HappyEngine/code/error/HeAssert.h
	src/HappyEngine/code/game/DynamicPhysicsComponent.cpp
	src/HappyEngine/code/game/DynamicPhysicsComponent.h
	src/HappyEngine/code/game/Entity.h
	src/HappyEngine/code/game/InstancedModelComponent.cpp
	src/HappyEngine/code/game/InstancedModelComponent.h
	src/HappyEngine/code/game/LightComponent.h
	src/HappyEngine/code/game/ModelComponent.cpp
	src/HappyEngine/code/game/ModelComponent.h
	src/HappyEngine/code/game/SkinnedModelComponent.cpp
	src/HappyEngine/code/game/StaticPhysicsComponent.h
	src/HappyEngine/code/game/TriggerComponent.h
	src/HappyEngine/code/graphics/AutoExposure.cpp
	src/HappyEngine/code/graphics/BillboardEffect.cpp
	src/HappyEngine/code/graphics/Bloom.cpp
	src/HappyEngine/code/graphics/BufferLayout.cpp
	src/HappyEngine/code/graphics/BufferLayout.h
	src/HappyEngine/code/graphics/DefaultSingleDrawable.cpp
	src/HappyEngine/code/graphics/DefaultSingleDrawable.h
	src/HappyEngine/code/graphics/DefaultSkinnedDrawable.cpp
	src/HappyEngine/code/graphics/DefaultSkinnedDrawable.h
	src/HappyEngine/code/graphics/Deferred3DRenderer.cpp
	src/HappyEngine/code/graphics/Deferred3DRenderer.h
	src/HappyEngine/code/graphics/DirectionalLight.h
	src/HappyEngine/code/graphics/DrawListContainer.cpp
	src/HappyEngine/code/graphics/DrawListContainer.h
	src/HappyEngine/code/graphics/Forward3DRenderer.cpp
	src/HappyEngine/code/graphics/IDrawable.h
	src/HappyEngine/code/graphics/InstancingController.cpp
	src/HappyEngine/code/graphics/InstancingController.h
	src/HappyEngine/code/graphics/Material.cpp
	src/HappyEngine/code/graphics/Material.h
	src/HappyEngine/code/graphics/Model.cpp
	src/HappyEngine/code/graphics/Model.h
	src/HappyEngine/code/graphics/ModelMesh.cpp
	src/HappyEngine/code/graphics/ModelMesh.h
	src/HappyEngine/code/graphics/PointLight.cpp
	src/HappyEngine/code/graphics/PointLight.h
	src/HappyEngine/code/graphics/PostProcesser.cpp
	src/HappyEngine/code/graphics/Shader.cpp
	src/HappyEngine/code/graphics/Shader.h
	src/HappyEngine/code/graphics/ShaderLayout.cpp
	src/HappyEngine/code/graphics/ShaderLayout.h
	src/HappyEngine/code/graphics/ShaderVar.h
	src/HappyEngine/code/graphics/Simple2DEffect.cpp
	src/HappyEngine/code/graphics/Simple2DTextureEffect.cpp
	src/HappyEngine/code/graphics/SimpleColorEffect.cpp
	src/HappyEngine/code/graphics/SkyBox.cpp
	src/HappyEngine/code/graphics/SkyBox.h
	src/HappyEngine/code/graphics/SpotLight.cpp
	src/HappyEngine/code/graphics/SpotLight.h
	src/HappyEngine/code/graphics/Texture2D.cpp
	src/HappyEngine/code/graphics/Texture2D.h
	src/HappyEngine/code/graphics/TextureCube.cpp
	src/HappyEngine/code/graphics/TextureCube.h
	src/HappyEngine/code/io/IniReader.cpp
	src/HappyEngine/code/io/IniReader.h
	src/HappyEngine/code/math/vec4.cpp
	src/HappyEngine/code/math/vec4.h

+++ src/HappyJson/src/HappyJson/json_valueiterator.inl
@@ -168,7 +168,7 @@ ValueIteratorBase::key() const
          return Value( StaticString( czstring.c_str() ) );
       return Value( czstring.c_str() );
    }
-   return Value( czstring.index() );
+   return Value( static_cast<UInt>(czstring.index()) );
 #else
    if ( isArray_ )
       return Value( ValueInternalArray::indexOf( iterator_.array_ ) );
